# Performance optimization for geometric kernel